[Tech] create hook to get vigilance areas and associate filters
